[Cleanup] Merge InspectorTreeControllerFlutter and superclass into single InspectorTreeController class (#3367)
